+++ FoundationSwaggerTests/Tools/FilePermissionsConstants.swift
@@ -1,0 +1,11 @@
+//
+//  FilePermissionsConstants.swift
+//  FoundationSwagger
+//
+//  Created by Sam Odom on 2/9/17.
+//  Copyright © 2017 Swagger Soft. All rights reserved.
+//
+
+import FoundationSwagger
+
+let MaximumFilePermissionsMask = FilePermissionsMask(UInt16(pow(2.0, 9)) - 1)